cloning: introduce 3 helper functions for CoA
The reference count should be record for each table not each entry.
The pud/pgd point to the same table (page), not an entry. Therefore,
we should record a table (page) is shared or not.

We used list for easy implementation, although list is not so efficient.
But I think it won't affect cloning, and we can consider a more efficient
to record sharing pfn (rb tree or something).