cloning: fix copy_to_user failed at a particular address
The root cause is the HVA in target VM is set as read-only, but
the address is set as read-write in source VM. At VM booting, the
first access to this address is *write*, therefore, write_fault in
user_mem_abort is true and cause gfn_to_pfn_prot() to allocate a
read-write entry.

However, the first access after cloning is *read*, it cause
gfn_to_pfn_prot() create a read-only page table entry, and
copy_to_user failed.

This is a work-around solution. We make every page table as read/write,
this is common, because hypervisor/qemu should has ability to read/write
all the guest memory. For stage2, it should also be read/write and leave
the permission to stage1 lookup.

The better solution would be postponed attributes modification until
target_copy_coa_page() and do some checks.