cloning: postpone RAM/IO address checking until stage2_set_pte
We postpone RAM/IO address checking until stage2_set_pte, because using
pgd index to distinguish is not enough, I/O address and RAM address
may belong to the same pgd entry.

In stage2_set_pte, we don't need to do coa_pte for such IO address, but
we still need to flush cache and remove the pfn from shared_list if the
pfn is existed.